file-upload: enable the file upload to continue when switching pages
